fix: used info
