--- src/components/DetailUser/utils.js
@@ -1,16 +1,0 @@
-export const getFormatedDate = dateString => {
-  const today = new Date(dateString);
-  const yyyy = today.getFullYear();
-  let mm = today.getMonth() + 1;
-  let dd = today.getDate();
-
-  if (dd < 10) {
-    dd = '0' + dd;
-  }
-
-  if (mm < 10) {
-    mm = '0' + mm;
-  }
-
-  return yyyy + '/' + mm + '/' + dd;
-};
