+++ src/components/Subtitle/Subtitle.module.css
@@ -1,6 +1,5 @@
 .text {
   margin-bottom: 32px;
-  color: var(--color-secondary);
 }
 
 /* tablet */
@@ -8,7 +7,6 @@
   .text {
     margin-bottom: 40px;
     max-width: 440px;
-    color: var(--color-text);
   }
 }
