+++ src/redux/store.js
@@ -6,7 +6,6 @@ import { ingredientsReducer } from './ingredients/slice';
 import { testimonialsReducer } from './testimonials/slice';
 import { commonReducer } from './common/slice';
 import { authReducer } from './auth/slice';
-import { usersReducer } from './users/slice';
 import storage from 'redux-persist/lib/storage';
 import { persistReducer, persistStore } from 'redux-persist';
 
@@ -26,7 +25,6 @@ export const store = configureStore({
     areas: areasReducer,
     ingredients: ingredientsReducer,
     testimonials: testimonialsReducer,
-    users: usersReducer,
   },
   middleware: getDefaultMiddleware =>
     getDefaultMiddleware({
