--- src/redux/users/selectors.js
@@ -1,4 +1,0 @@
-export const selectUsers = state => state.users.items;
-export const selectCurrentUser = state => state.users.user;
-export const selectLoading = state => state.users.loading;
-export const selectError = state => state.users.error;
